Remember the "found declaration" for an overload candidate, which is the entity (if applicable) which was actually looked up.  If a candidate was found via a using declaration, this is the UsingShadowDecl;  otherwise, if the candidate is template specialization, this is the template;  otherwise, this is the function.
The point of this exercise is that "found declarations" are the entities
we do access control for, not their underlying declarations.  Broadly speaking,
this patch fixes access control for using declarations.

There is a *lot* of redundant code calling into the overload-resolution APIs;
we really ought to clean that up.



git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@98945 91177308-0d34-0410-b5e6-96231b3b80d8