Merging r143345: ------------------------------------------------------------------------ r143345 | chandlerc | 2011-10-31 01:42:27 -0700 (Mon, 31 Oct 2011) | 7 lines
Add "just one more" include path to the monstrosity that is our header
search logic. The Debian multiarch seems to have completely changed from
when these were originally added, and I'd like to remove a bunch of
them, but I'll be lazy and delay that until this logic is hoisted into
the driver where it belongs.

This should resolve PR11223.
------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/cfe/branches/release_30@143433 91177308-0d34-0410-b5e6-96231b3b80d8