Fix up the BUILD_CLANG_ONLY code to work properly after r184794.
When BUILD_CLANG_ONLY is set to YES, it is supposed to simply limit the tools
that get built.  The change in r184794 broke this feature by moving libclang
and c-index-test into PARALLEL_DIRS. Those are both supposed to be in DIRS,
because c-index-test has a build dependency on libclang and cannot be
reliably built in parallel with it.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@187246 91177308-0d34-0410-b5e6-96231b3b80d8