Use backslashes to escape spaces and other backslashes in -dwarf-debug-flags.
The command line options are specified in a space-separated list that is an
argument to -dwarf-debug-flags, so that breaks if there are spaces in the
options. This feature came from Apple's internal version of GCC, so I went back
to check how llvm-gcc handled this and matched that behavior.
rdar://problem/18775420

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@221309 91177308-0d34-0410-b5e6-96231b3b80d8