Wdeprecated: Make BumpVectorContext movable
Turns out the one place that relied on the implicit copy ctor was safe
because it created an object in a state where the dtor was a no-op, but
that's more luck that good management.

Sure up the API by defining move construction and using it, which
implicitly disallows the unreliable copy operations.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@244968 91177308-0d34-0410-b5e6-96231b3b80d8