Merging r245041: ------------------------------------------------------------------------ r245041 | abataev | 2015-08-14 05:25:37 -0700 (Fri, 14 Aug 2015) | 4 lines
[OPENMP] Fix for http://llvm.org/PR24371: Assert failure compiling blender 2.75.
blender uses statements expression in condition of the loop under control of the '#pragma omp parallel for'. This condition is used several times in different expressions required for codegen of the loop directive. If there are some variables defined in statement expression, it fires an assert during codegen because of redefinition of the same variables.
We have to rebuild several expression to be sure that all variables are unique.

------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/cfe/branches/release_37@245229 91177308-0d34-0410-b5e6-96231b3b80d8