Link include-fixer into libclang if clang-tools-extra is checked out.
include-fixer only slightly bloats the size of libclang, but since
libclang has no explicit plugin mechanism it's the only way of getting
this to work. Clang-tidy is already there and so far there weren't many
complaints ;)

This is designed to be easy to remove again if libclang ever grows
proper plugin support.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@287229 91177308-0d34-0410-b5e6-96231b3b80d8