Make ASTContext::getDeclAlign return the correct alignment for FunctionDecls.
This commit silences an incorrect warning that is issued when a function
pointer is cast to another function pointer type. The warning gets
issued because alignments of the source and destination do not match in
Sema::CheckCastAlign, which happens because ASTContext::getTypeInfoImpl
and ASTContext::getDeclAlign return different values for functions (the
former returns 4 while the latter returns 1).

This should fix PR31558.

rdar://problem/29533528

Differential Revision: https://reviews.llvm.org/D27478


git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@291253 91177308-0d34-0410-b5e6-96231b3b80d8