[X86] Remove 'umip' feature flag.
This feature flag indicates that the processor has support for removing certain instructions from user mode software. But the feature flag by itself doesn't indicate if the support is enabled in the OS. The affected instructions aren't even instructions the compiler would emit. So I don't think think this feature flag should be in the compiler.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@294414 91177308-0d34-0410-b5e6-96231b3b80d8