Partially fix comment in test broken in r306079 and r306948
A RUN line was referring to the previous RUN line but a new test was added in
between them.  Just reorder the lines.

Note this still does not completely fix this the brokenness of the comment as
the driver-based test gained a new hotness-threshold argument in r306948 but
I'll fix that is a separate commit.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@319576 91177308-0d34-0410-b5e6-96231b3b80d8