[X86] Remove 'rtm' feature from KNL.
I'm unsure if KNL has this feature, but the backend never thought it did, only clang did. The predefined-arch-macros test lost the check for __RTM__ on KNL when it was removed Skylake CPUs in r344117.

I think we want to drop it from KNL for consistency with Skylake anyway regardless of how we got here.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@344978 91177308-0d34-0410-b5e6-96231b3b80d8