Dump actual line numbers when dumping the AST to JSON.
The "line" attribute is now the physical line within the source file for the location. A "presumedLine" attribute is printed when the presumed line number does not match the given source line number. We continue to not print repeated line information in subsequent source locations, but we track presumed and actual lines separately.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@365919 91177308-0d34-0410-b5e6-96231b3b80d8