[CFG] Fix CFG for statement-expressions in return values.
We're building the CFG from bottom to top, so when the return-value expression
has a non-trivial CFG on its own, we need to continue building from the entry
to the return-value expression CFG rather than from the block to which
we've just appended the return statement.

Fixes a false positive warning "control may reach end of non-void function".

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@370406 91177308-0d34-0410-b5e6-96231b3b80d8